Moves common Module form clean logic to new class
